Fork socket server process and fire it up from main index
