+++ main/server/messageTypes.js
@@ -1,0 +1,1 @@
+export const PING = 'PING'
