+++ webpack.config.js
@@ -251,4 +251,49 @@ const loginPopupConfig = {
   optimization: { splitChunks: false },
 }
 
-module.exports = [rendererConfig, mainConfig, loginPopupConfig]
+const socketServerConfig = {
+  mode: process.env.NODE_ENV === 'dev' ? 'development' : 'production',
+  watch: process.env.NODE_ENV === 'dev',
+  context: path.resolve(__dirname, 'main'),
+  entry: {
+    socketServer: path.resolve('.', 'main', 'server', 'server.js'),
+  },
+  output: {
+    path: isForMaps ? sourceMapsPath : path.resolve(__dirname, 'bin'),
+    filename: '[name].bundle.js',
+  },
+  module: {
+    rules: [
+      {
+        test: /\.js$/,
+        loader: 'babel-loader',
+        include: path.resolve(__dirname, 'lib', 'pltr'),
+        exclude: /node_modules/,
+      },
+      {
+        test: /\.js$/,
+        loader: 'babel-loader',
+        include: path.resolve(__dirname, 'src'),
+        exclude: /node_modules/,
+      },
+    ],
+  },
+  resolve: {
+    extensions: ['.js', '.json'],
+    modules: ['main', 'node_modules'],
+    alias: {
+      'connected-components': path.resolve(__dirname, 'src', 'connected-components.js'),
+      'wired-up-firebase': path.resolve(__dirname, 'src', 'wired-up-firebase.js'),
+      'world-api': path.resolve(__dirname, 'src', 'world.js'),
+      plottr_components: path.resolve(__dirname, 'lib', 'plottr_components', 'dist', 'components'),
+      docx: path.resolve('./node_modules/docx'),
+      plottr_locales: path.resolve('./node_modules/plottr_locales'),
+    },
+  },
+  target: 'electron-renderer',
+  plugins: [appCircularDependencyChecker, duplicateDependencyChecker, ...plugins],
+  devtool: process.env.NODE_ENV === 'dev' ? 'eval' : false,
+  optimization: { splitChunks: false },
+}
+
+module.exports = [rendererConfig, mainConfig, loginPopupConfig, socketServerConfig]
